# clients/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "holm",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/boxes/{id}/cdp": {
      "post": {
        "tags": [
          "pages"
        ],
        "operationId": "token",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ttl_secs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Done",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CdpToken"
                }
              }
            }
          },
          "default": {
            "description": "The failure, as an ErrorBody",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CdpToken": {
        "type": "object",
        "required": [
          "url",
          "ws_url",
          "expires_at_ms"
        ],
        "properties": {
          "expires_at_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "url": {
            "type": "string"
          },
          "ws_url": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "bad_request",
          "not_found",
          "gone",
          "denied",
          "screen_unavailable",
          "unsupported",
          "failed",
          "timeout",
          "unavailable",
          "transport",
          "internal"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}